i101: Split map content and persist local shouts
