+++ react-app/src/screens/HomeScreen.js
@@ -8,7 +8,7 @@ import { normalizeStyles } from 'res';
 const HomeScreen = ({ storageLoaded, navigation }) => {
   const { STRINGS } = useContext(Config.Context);
   return (
-    <World>
+    <World onboarding={false} >
       {storageLoaded &&
       <View style={styles.container} pointerEvents='box-none'>
         <HomeBar />
